Testing fixed stepSize variable. Seems to be working

+++ Task61.h
@@ -62,7 +62,7 @@ public:
 	//Values for LIC
 	int SampleX;
 	int SampleY;
-	int KernelSize;
+	float KernelSize;
 	int Seed;
 	bool RandomTexture;
 	bool FastLIC;
